+++ sdk/core/azure-core/CHANGELOG.md
@@ -17,6 +17,7 @@
 - Renamed `Azure::Nullable<T>::GetValue()` to `Value()`.
 - Removed from `Azure::Core::Http::Policies` namespace: `HttpPolicyOrder`, `TransportPolicy`, `RetryPolicy`, `RequestIdPolicy`, `TelemetryPolicy`, `BearerTokenAuthenticationPolicy`, `LogPolicy`.
 - Renamed `Azure::Core::Http::RawResponse::GetBodyStream()` to `ExtractBodyStream()`.
+- Introduced `Azure::Core::Context::Key` class which takes place of `std::string` used for `Azure::Core::Context` keys previously.
 
 ## 1.0.0-beta.7 (2021-03-11)